Fixing compilation issues under Arduino IDE (w/ IDF 5).

+++ src/roo_logging/stream.cpp
@@ -59,7 +59,7 @@ DefaultLogStream& operator<<(DefaultLogStream& s, roo_time::Interval::Components
   bool force = false;
   if (components.negative) s << "-";
   if (components.days > 0) {
-    s.printf("%d.", components.days);
+    s.printf("%d.", (int)components.days);
     force = true;
   }
   if (force || components.hours > 0) {
